Fixes based on comments regarding building on windows.
Specifically, python 3.9 and newer does not consume quotes on the
command line, so additional escaping is not needed.

Secondly, openssl lib name may be libcrypto.lib on newer openssl
versions, so I've added an option to specify this with an environment
variable OPENSSL_LIBNAME (which defaults to libeay32.lib for now).